requirements.txt
katna
moviepy
openai
pysrt
toml
pillow
numpy
re
python-docx